flask_app/backend/authentication_service/requirements.txt
anyio==4.2.0
ariadne==0.21
bidict==0.22.1
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-engineio==4.8.1
python-socketio==5.10.0
simple-websocket==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.24
starlette==0.34.0
typing_extensions==4.9.0
Werkzeug==3.0.1
wsproto==1.2.0
